gateway/requirements.txt
# HTTP client
requests>=2.28.2
websockets>=10.4

# Command line parsing
argparse>=1.4.0

# Logging
logging>=0.4.9

# Date and time handling
python-dateutil>=2.8.2

# Async programming
asyncio>=3.4.3

# File system monitoring
watchdog>=2.1.9

# Serial port communication
pyserial>=3.5
pyserial-asyncio>=0.6

# Environment variables
python-dotenv>=0.21.0
